requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=1.6
SQLAlchemy<1.1.0,>=0.9.9
boto>=2.32.1
decorator>=3.4.0
eventlet>=0.17.4
Jinja2>=2.6 # BSD License (3 clause)
keystonemiddleware!=2.4.0,>=2.0.0
lxml>=2.3
Routes!=2.0,!=2.1,>=1.12.3;python_version=='2.7'
Routes!=2.0,>=1.12.3;python_version!='2.7'
cryptography>=1.0 # Apache-2.0
WebOb>=1.2.3
greenlet>=0.3.2
PasteDeploy>=1.5.0
Paste
PrettyTable<0.8,>=0.7
sqlalchemy-migrate>=0.9.6
netaddr!=0.7.16,>=0.7.12
netifaces>=0.10.4
paramiko>=1.13.0
Babel>=1.3
iso8601>=0.1.9
jsonschema!=2.5.0,<3.0.0,>=2.0.0
python-cinderclient>=1.3.1
python-keystoneclient!=1.8.0,>=1.6.0
python-neutronclient>=2.6.0
python-glanceclient>=0.18.0
requests!=2.8.0,>=2.5.2
six>=1.9.0
stevedore>=1.5.0 # Apache-2.0
setuptools>=16.0
websockify>=0.6.1
oslo.concurrency>=2.3.0 # Apache-2.0
oslo.config>=2.6.0 # Apache-2.0
oslo.context>=0.2.0 # Apache-2.0
oslo.log>=1.8.0 # Apache-2.0
oslo.reports>=0.6.0 # Apache-2.0
oslo.serialization>=1.10.0 # Apache-2.0
oslo.utils!=2.6.0,>=2.4.0 # Apache-2.0
oslo.db>=3.0.0 # Apache-2.0
oslo.rootwrap>=2.0.0 # Apache-2.0
oslo.messaging!=1.17.0,!=1.17.1,!=2.6.0,!=2.6.1,>=1.16.0 # Apache-2.0
oslo.i18n>=1.5.0 # Apache-2.0
oslo.service>=0.10.0 # Apache-2.0
rfc3986>=0.2.0 # Apache-2.0
oslo.middleware>=2.8.0 # Apache-2.0
psutil<2.0.0,>=1.1.1
oslo.versionedobjects>=0.9.0
alembic>=0.8.0
os-brick>=0.4.0 # Apache-2.0
logbook
pyzmq
